Revert "graphile-config/helpers and beginning of lib system"
This reverts commit 57df844f72c40e58a9bc9972408b7b0a2f7e82c0.
