+++ utils/graphile-config/package.json
@@ -14,10 +14,6 @@
       "types": "./dist/loadConfig.d.ts",
       "default": "./dist/loadConfig.js"
     },
-    "./helpers": {
-      "types": "./dist/helpers.d.ts",
-      "default": "./dist/helpers.js"
-    },
     "./cli": {
       "types": "./dist/cli.d.ts",
       "default": "./dist/cli.js"
