+++ utils/graphile-config/src/resolvePresets.ts
@@ -37,9 +37,9 @@ try {
   };
 }
 
-export function isResolvedPreset<TLib extends Record<string, unknown>>(
-  preset: GraphileConfig.Preset<TLib, any>,
-): preset is GraphileConfig.ResolvedPreset<TLib> {
+export function isResolvedPreset(
+  preset: GraphileConfig.Preset,
+): preset is GraphileConfig.ResolvedPreset {
   return (
     (preset.plugins &&
       preset.extends?.length === 0 &&
@@ -52,48 +52,21 @@ export function isResolvedPreset<TLib extends Record<string, unknown>>(
   );
 }
 
-/** @deprecated Use `resolvePreset({ extends: presets })` instead */
-export function resolvePresets(
-  presets: ReadonlyArray<GraphileConfig.Preset<any, any>>,
-) {
-  if (presets.length === 1) {
-    return resolvePreset(presets[0]);
-  }
-  return resolvePreset({ extends: presets });
-}
-
-type UnionToIntersection<U> = (U extends any ? (k: U) => void : never) extends (
-  k: infer I,
-) => void
-  ? I
-  : never;
-export type Simplify<T> = {} & { [P in keyof T]: T[P] };
-
 /**
- * Given a preset, recursively walks the `extends` and returns the resulting
+ * Given a list of presets, resolves the presets and returns the resulting
  * ResolvedPreset (which does not have any `extends`).
  */
-export function resolvePreset<
-  TLib extends Record<string, unknown>,
-  TExtends extends ReadonlyArray<GraphileConfig.Preset<any, any>>,
->(
-  preset: GraphileConfig.Preset<TLib, TExtends>,
-): GraphileConfig.ResolvedPreset<
-  Simplify<
-    TLib & TExtends extends readonly GraphileConfig.Preset<infer U, any>[]
-      ? UnionToIntersection<any extends U ? {} : U>
-      : {}
-  >
-> {
-  if (typeof preset !== "object" || preset === null) {
-    throw new Error(`Expected a preset, received ${inspect(preset)}`);
+export function resolvePresets(presets: ReadonlyArray<GraphileConfig.Preset>) {
+  if (presets.length === 1) {
+    // Maybe it's already resolved?
+    const preset = presets[0];
+    if (preset && isResolvedPreset(preset)) {
+      return preset;
+    }
   }
-  // Maybe it's already resolved?
-  if (isResolvedPreset(preset)) {
-    return preset as any;
-  }
+
   const seenPluginNames = new Set<string>();
-  const resolvedPreset = resolvePresetsInternal([preset], seenPluginNames, 0);
+  const resolvedPreset = resolvePresetsInternal(presets, seenPluginNames, 0);
 
   const disabledButNotSeen = resolvedPreset.disablePlugins?.filter(
     (n) => !seenPluginNames.has(n),
@@ -112,15 +85,10 @@ export function resolvePreset<
 }
 
 function resolvePresetsInternal(
-  presets: ReadonlyArray<
-    GraphileConfig.Preset<
-      Record<string, any>,
-      readonly GraphileConfig.Preset<any, any>[]
-    >
-  >,
+  presets: ReadonlyArray<GraphileConfig.Preset>,
   seenPluginNames: Set<string>,
   depth: number,
-): GraphileConfig.ResolvedPreset<any> {
+): GraphileConfig.ResolvedPreset {
   const finalPreset = blankResolvedPreset();
   for (const preset of presets) {
     if (isResolvedPreset(preset) && preset.disablePlugins) {
